shaders: add drawing system with handle registry
