+++ shaders/__init__.py
@@ -1,0 +1,3 @@
+from . import handle
+
+__all__ = ["handle"]
